This is the final paper for the summer research
The paper focuses on a Faster Solver of Multiple Linear Systems via Block Conjugate Gradients. We compare equations that describe the time and memory costs for several analogous algorithms to BCG. We plot graphs for convergence and time spent on these algorithms. We also find out another method to avoid the singular matrix issue that we encounter in the block version algorithms.